Serialize(?) test invocations to see if that fixes bizarre CI errors
