+++ cargo-audit/tests/acceptance.rs
@@ -28,7 +28,6 @@ static ADVISORY_DB_DIR: Lazy<TempDir> = Lazy::new(|| TempDir::new().unwrap());
 pub static RUNNER: Lazy<CmdRunner> = Lazy::new(|| {
     let mut runner = CmdRunner::default();
     runner.arg("audit").arg("--db").arg(ADVISORY_DB_DIR.path());
-    runner.capture_stdout().capture_stderr();
     runner
 });
 
@@ -171,4 +170,40 @@ fn advisories_found_but_ignored_json() {
             .unwrap(),
         0
     );
+}
+
+#[cfg(feature = "binary-scanning")]
+fn binaries_dir() -> PathBuf {
+    [env!("CARGO_MANIFEST_DIR"), "tests", "support", "binaries"]
+        .iter()
+        .collect()
+}
+
+#[cfg(feature = "binary-scanning")]
+fn bin_cmd_runner() -> CmdRunner {
+    RUNNER.clone()
+}
+
+#[cfg(feature = "binary-scanning")]
+#[test]
+fn binary_without_audit_info_is_rejected() {
+    let mut binary_path = binaries_dir();
+    binary_path.push("binary-without-audit-info");
+    assert_eq!(bin_cmd_runner().arg(binary_path).status().code(), 2);
+}
+
+#[cfg(feature = "binary-scanning")]
+#[test]
+fn binary_without_vulnerabilities_passes() {
+    let mut binary_path = binaries_dir();
+    binary_path.push("binary-with-audit-info");
+    assert_eq!(bin_cmd_runner().arg(binary_path).status().code(), 0);
+}
+
+#[cfg(feature = "binary-scanning")]
+#[test]
+fn binary_with_vulnerabilities_fails() {
+    let mut binary_path = binaries_dir();
+    binary_path.push("binary-with-vuln");
+    assert_eq!(bin_cmd_runner().arg(binary_path).status().code(), 1);
 }
